requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.0.1
efficientnet-pytorch==0.7.1
torchvision==0.15.2
pillow==10.0.0
pdf2image==1.16.3